+++ models/moving_average.rb
@@ -1,4 +1,4 @@
-class MovingAverage < ActiveRecord::Base
+class MovingAverage < ApplicationRecord
   PAIR_LIST = %w[AUDJPY CADJPY CHFJPY EURJPY EURUSD GBPJPY NZDJPY USDJPY].freeze
   TIME_FRAME_LIST = %w[M1 M5 M15 M30 H1 H4 D1 W1 MN1].freeze
   PERIOD_LIST = [25, 50, 75, 100, 150, 200].freeze
@@ -13,4 +13,18 @@ class MovingAverage < ActiveRecord::Base
             inclusion: {in: PERIOD_LIST, message: 'invalid'}
   validates :value,
             numericality: {greater_than: 0, message: 'invalid'}
+
+  scope :on, lambda {|date|
+    from = date.strftime('%F 00:00:00')
+    to = date.strftime('%F 23:59:59')
+    where('`time` BETWEEN ? AND ?', from, to)
+  }
+
+  def self.create_infile(src_file, dst_file)
+    FileUtils.cp(src_file, dst_file)
+  end
+
+  def to_csv
+    [time.strftime('%F %T'), pair, time_frame, period, value]
+  end
 end